[#200] :recycle: refactor: api, form api instance 파일 분리
